updated with fishnix_s PR comments
